move Exception class
