+++ resources/views/cms/layout/templates/sidebar/shortcuts.blade.php
