+++ resources/views/cms/pages/administration/templates/edit-info.blade.php
