+++ resources/views/cms/pages/configuration/applications.blade.php
