+++ resources/views/cms/pages/content/assign.blade.php
