+++ resources/views/cms/pages/content/create.blade.php
