+++ resources/views/cms/pages/content/home.blade.php
