+++ resources/views/cms/pages/content/index.blade.php
