+++ resources/views/cms/pages/content/templates/contentController.blade.php
