+++ resources/views/cms/pages/post/create.blade.php
