+++ resources/views/cms/pages/post/index.blade.php
