+++ resources/views/cms/pages/post/show.blade.php
