+++ resources/views/emails/auth/reminder.blade.php
