+++ resources/views/emails/cms/resetPassword.blade.php
